# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
...
      "BulkResult": {
        "properties": {
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/RoomOut"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "created",
          "data"
        ],
        "title": "BulkResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportReport": {
        "properties": {
          "committed": {
            "title": "Committed",
            "type": "boolean"
          },
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ImportProblem"
            },
            "title": "Errors",
            "type": "array"
          },
          "rows": {
            "title": "Rows",
            "type": "integer"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "committed",
          "valid",
          "rows",
          "errors",
          "created"
        ],
        "title": "ImportReport",
        "type": "object"
      },
...
      "Money": {
        "additionalProperties": false,
        "properties": {
          "amount_minor": {
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "amount_minor",
          "currency"
        ],
        "title": "Money",
        "type": "object"
      },
...
      "RoomBulkCreate": {
        "additionalProperties": false,
        "properties": {
          "ranges": {
            "items": {
              "$ref": "#/components/schemas/RoomRange"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Ranges",
            "type": "array"
          }
        },
        "required": [
          "ranges"
        ],
        "title": "RoomBulkCreate",
        "type": "object"
      },
      "RoomCreate": {
        "additionalProperties": false,
        "properties": {
          "amenities": {
            "items": {
              "maxLength": 60,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 50,
            "title": "Amenities",
            "type": "array"
          },
          "capacity": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "floor": {
            "anyOf": [
              {
                "maxLength": 20,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "number": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Number",
            "type": "string"
          },
          "rate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "room_type_id": {
            "format": "uuid",
            "title": "Room Type Id",
            "type": "string"
          }
        },
        "required": [
          "number",
          "room_type_id"
        ],
        "title": "RoomCreate",
        "type": "object"
      },
      "RoomList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RoomOut"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "RoomList",
        "type": "object"
      },
      "RoomOut": {
        "properties": {
          "amenities": {
            "items": {
              "type": "string"
            },
            "title": "Amenities",
            "type": "array"
          },
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "effective_rate": {
            "$ref": "#/components/schemas/Money"
          },
          "floor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "rate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "room_type": {
            "$ref": "#/components/schemas/RoomTypeRef"
          },
          "status": {
            "enum": [
              "available",
              "occupied",
              "reserved",
              "cleaning",
              "maintenance",
              "out_of_service"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_changed_at": {
            "format": "date-time",
            "title": "Status Changed At",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "number",
          "floor",
          "room_type",
          "rate",
          "effective_rate",
          "capacity",
          "amenities",
          "status",
          "status_changed_at",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "RoomOut",
        "type": "object"
      },
      "RoomPatch": {
        "additionalProperties": false,
        "properties": {
          "amenities": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 60,
                  "minLength": 1,
                  "type": "string"
                },
                "maxItems": 50,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenities"
          },
          "capacity": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "floor": {
            "anyOf": [
              {
                "maxLength": 20,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "number": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "rate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "room_type_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Type Id"
          }
        },
        "title": "RoomPatch",
        "type": "object"
      },
...
      "RoomStatusChange": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "enum": [
              "available",
              "reserved",
              "cleaning",
              "maintenance",
              "out_of_service"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "RoomStatusChange",
        "type": "object"
      },
      "RoomTypeCreate": {
        "additionalProperties": false,
        "properties": {
          "amenities": {
            "items": {
              "maxLength": 60,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 50,
            "title": "Amenities",
            "type": "array"
          },
          "base_rate": {
            "$ref": "#/components/schemas/Money"
          },
          "capacity": {
            "default": 2,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Capacity",
            "type": "integer"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "base_rate"
        ],
        "title": "RoomTypeCreate",
        "type": "object"
      },
      "RoomTypeList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RoomTypeOut"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "RoomTypeList",
        "type": "object"
      },
      "RoomTypeOut": {
        "properties": {
          "amenities": {
            "items": {
              "type": "string"
            },
            "title": "Amenities",
            "type": "array"
          },
          "base_rate": {
            "$ref": "#/components/schemas/Money"
          },
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "room_count": {
            "title": "Room Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "base_rate",
          "capacity",
          "amenities",
          "room_count",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "RoomTypeOut",
        "type": "object"
      },
      "RoomTypePatch": {
        "additionalProperties": false,
        "properties": {
          "amenities": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 60,
                  "minLength": 1,
                  "type": "string"
                },
                "maxItems": 50,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenities"
          },
          "base_rate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "capacity": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "RoomTypePatch",
        "type": "object"
      },
      "RoomTypeRef": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "RoomTypeRef",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Diyneco API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/room-types": {
      "get": {
        "operationId": "list_room_types_api_v1_room_types_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomTypeList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Room Types",
        "tags": [
          "rooms"
        ]
      },
      "post": {
        "operationId": "create_room_type_api_v1_room_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomTypeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomTypeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Room Type",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/room-types/{room_type_id}": {
      "patch": {
        "operationId": "patch_room_type_api_v1_room_types__room_type_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "room_type_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Room Type Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomTypePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomTypeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Room Type",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms": {
      "get": {
        "operationId": "list_rooms_api_v1_rooms_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "available",
                    "occupied",
                    "reserved",
                    "cleaning",
                    "maintenance",
                    "out_of_service"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "floor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Floor"
            }
          },
          {
            "in": "query",
            "name": "room_type_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Room Type Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rooms",
        "tags": [
          "rooms"
        ]
      },
      "post": {
        "operationId": "create_room_api_v1_rooms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Room",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/bulk": {
      "post": {
        "operationId": "bulk_create_rooms_api_v1_rooms_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomBulkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Create Rooms",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/import": {
      "post": {
        "description": "Dry run by default: returns a per-line report. `?commit=true` writes all rows or none.",
        "operationId": "import_rooms_api_v1_rooms_import_post",
        "parameters": [
          {
            "in": "query",
            "name": "commit",
            "required": false,
            "schema": {
              "default": false,
              "title": "Commit",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/csv": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Rooms",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{room_id}": {
      "delete": {
        "operationId": "delete_room_api_v1_rooms__room_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Room",
        "tags": [
          "rooms"
        ]
      },
      "get": {
        "operationId": "get_room_api_v1_rooms__room_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Room",
        "tags": [
          "rooms"
        ]
      },
      "patch": {
        "operationId": "patch_room_api_v1_rooms__room_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Room Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Room",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{room_id}/status": {
      "post": {
        "operationId": "set_room_status_api_v1_rooms__room_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomStatusChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Room Status",
        "tags": [
          "rooms"
        ]
      }
    },
...
  }
}